CALY UKLAD + KULKA
wypakuj znowu pliki obj bo poprawiłem kulke

+++ Arkanoid3D/Ball.h
@@ -19,11 +19,11 @@ public:
 	Ball();
 	void drawBall();
 
-	int getX() {
+	float getX() {
 		return this->X;
 	}
 
-	int getY() {
+	float getY() {
 		return this->Y;
 	}
 };
